requirements.txt
Django==2.0.3
django-extensions==2.0.6
django-forms-bootstrap==3.1.0
django-multiupload==0.5.2
pytz==2018.3
six==1.11.0
